determinism fixes for gradient registration 1. Converting metric values to BigDecimal for stable parallel summation 2. Avoid using caching in SurfaceSpatialIndex as this results in different values for the same point when ran in parallel
Additionally build.sbt contains fixes as v0.90.0 would not compile

+++ src/main/scala/scalismo/registration/MeanPointwiseLossMetric.scala
@@ -78,7 +78,7 @@ abstract class MeanPointwiseLossMetric[D: NDSpace, A: Scalar](
 
     // we compute the mean using a monte carlo integration
     val samples = sampler.sample()
-    new ParVector(samples.toVector).map { case (pt, _) => metricValue(pt).getOrElse(0.0) }.sum / samples.size
+    (new ParVector(samples.toVector).map { case (pt, _) => BigDecimal(metricValue(pt).getOrElse(0.0)) }.sum / samples.size).toDouble
   }
 
   private def computeDerivative(parameters: DenseVector[Double], sampler: Sampler[D]): DenseVector[Double] = {
